# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_leleo886_lopic_internal_error.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  services.AlbumResponse:
    properties:
      cover_image:
        type: string
      created_at:
        type: string
      description:
        type: string
      gallery_enabled:
        type: boolean
      id:
        type: integer
      image_count:
        type: integer
      name:
        type: string
      serial_number:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  services.GetAlbumsResponse:
    properties:
      albums:
        items:
          $ref: '#/definitions/services.AlbumResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  success.DataResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  success.SuccessResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: RESTful API
  title: Lopic API
  version: 1.0.0
paths:
  /api/admin/albums:
    delete:
      description: 删除指定相册
      parameters:
      - description: 相册ID列表
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/success.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除相册，只删除相册与图片的关联信息，不删除图片
      tags:
      - 相册管理员
    get:
      description: 获取所有相册，支持分页
      parameters:
      - description: 页码，默认为1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认为10
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词
        in: query
        name: searchkey
        type: string
      - description: 排序字段，默认为created_at
        in: query
        name: orderby
        type: string
      - description: 排序方向，asc或desc，默认为desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.GetAlbumsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取所有相册列表
      tags:
      - 相册管理员
  /api/admin/albums/{id}:
    get:
      description: 获取指定相册的详细信息
      parameters:
      - description: 相册ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.AlbumResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取相册详情
      tags:
      - 相册管理员
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"